requirements.txt
appdirs==1.4.3
asgiref==3.4.1
attrs==19.1.0
black==19.3b0
Click==7.0
dataclasses==0.6
Django==3.1.12
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.8
marshmallow==3.2.0
mccabe==0.6.1
pycodestyle==2.5.0
pydantic==1.6.2
pyflakes==2.1.1
pytz==2019.1
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.3
typesystem==0.2.4
typing-extensions==3.7.4
